Introduce shared "draw two instead of one for additional draws" effect
