+++ Mage.Sets/src/mage/cards/b/BardKingOfDale.java
@@ -4,22 +4,15 @@ import java.util.UUID;
 import mage.MageInt;
 import mage.constants.SubType;
 import mage.constants.SuperType;
-import mage.game.Game;
-import mage.game.events.GameEvent;
-import mage.players.Player;
 import mage.watchers.common.CardsDrawnDuringDrawStepWatcher;
-import mage.abilities.Ability;
 import mage.abilities.common.SimpleStaticAbility;
-import mage.abilities.effects.ReplacementEffectImpl;
+import mage.abilities.effects.common.replacement.DrawExceptFirstDrawTwoReplacementEffect;
 import mage.abilities.effects.common.replacement.CreateTwiceThatManyTokensEffect;
 import mage.abilities.keyword.ReachAbility;
 import mage.abilities.keyword.VigilanceAbility;
 import mage.cards.CardImpl;
 import mage.cards.CardSetInfo;
 import mage.constants.CardType;
-import mage.constants.Duration;
-import mage.constants.Outcome;
-import mage.constants.PhaseStep;
 
 /**
  *
@@ -44,7 +37,7 @@ public final class BardKingOfDale extends CardImpl {
         this.addAbility(VigilanceAbility.getInstance());
 
         // If you would draw a card except the first one you draw in each of your draw steps, draw two cards instead.
-        this.addAbility(new SimpleStaticAbility(new BardKingOfDaleReplacementEffect()), new CardsDrawnDuringDrawStepWatcher());
+        this.addAbility(new SimpleStaticAbility(new DrawExceptFirstDrawTwoReplacementEffect()), new CardsDrawnDuringDrawStepWatcher());
 
         // If one or more tokens would be created under your control, twice that many of those tokens are created instead.
         this.addAbility(new SimpleStaticAbility(new CreateTwiceThatManyTokensEffect()));
@@ -57,49 +50,5 @@ public final class BardKingOfDale extends CardImpl {
     @Override
     public BardKingOfDale copy() {
         return new BardKingOfDale(this);
-    }
-}
-
-class BardKingOfDaleReplacementEffect extends ReplacementEffectImpl {
-
-    BardKingOfDaleReplacementEffect() {
-        super(Duration.WhileOnBattlefield, Outcome.Neutral);
-        staticText = "If you would draw a card except the first one you draw in each of your draw steps, draw two cards instead";
-    }
-
-    private BardKingOfDaleReplacementEffect(final BardKingOfDaleReplacementEffect effect) {
-        super(effect);
-    }
-
-    @Override
-    public BardKingOfDaleReplacementEffect copy() {
-        return new BardKingOfDaleReplacementEffect(this);
-    }
-
-    @Override
-    public boolean replaceEvent(GameEvent event, Ability source, Game game) {
-        Player controller = game.getPlayer(source.getControllerId());
-        if (controller != null) {
-            controller.drawCards(2, source, game, event);
-        }
-        return true;
-    }
-
-    @Override
-    public boolean checksEventType(GameEvent event, Game game) {
-        return event.getType() == GameEvent.EventType.DRAW_CARD;
-    }
-
-    @Override
-    public boolean applies(GameEvent event, Ability source, Game game) {
-        if (!event.getPlayerId().equals(source.getControllerId())) {
-            return false;
-        }
-        if (!game.isActivePlayer(event.getPlayerId())
-                || game.getPhase().getStep().getType() != PhaseStep.DRAW) {
-            return true;
-        }
-        CardsDrawnDuringDrawStepWatcher watcher = game.getState().getWatcher(CardsDrawnDuringDrawStepWatcher.class);
-        return watcher != null && watcher.getAmountCardsDrawn(event.getPlayerId()) > 0;
     }
 }
